Split taxes module into submodules
